refactor(types): organize types
