+++ app/components/TotalDeposit.tsx
@@ -2,9 +2,7 @@ import { Info, Zap } from "lucide-react";
 import { Card, CardContent } from "./ui/card";
 import { TooltipContent, TooltipTrigger } from "./ui/tooltip";
 import { Table } from "./ui/table";
-import { vaults } from "~/constant";
 import { Column, CellProps } from "react-table";
-import { DApp } from "~/types";
 import { Badge } from "./ui/badge";
 import { Link } from "@remix-run/react";
 import { SelectInput } from "./ui/select";
@@ -17,26 +15,14 @@ interface DepositData {
 	tooltip: string;
 }
 
-const MOCK_DEPOSIT_DATA: DepositData[] = [
-	{
-		id: "totalValue",
-		title: "Total Value Routed",
-		value: "$9.62M",
-		tooltip: "",
-	},
-	{
-		id: "totalDeposit",
-		title: "Total Deposit",
-		value: "$14.72M",
-		tooltip: "",
-	},
-	{
-		id: "totalBorrow",
-		title: "Total Borrow",
-		value: "$9.23M",
-		tooltip: "",
-	},
-];
+interface DApp {
+	name: string;
+	type: string;
+	labels: string[];
+	apy: number;
+	chain: string;
+	logo: string;
+}
 
 export const columns: Column<DApp>[] = [
 	{
@@ -138,8 +124,96 @@ export function TotalDeposit() {
 						<Input type="text" placeholder="Search vaults..." />
 					</div>
 				</div>
-				<Table columns={columns} data={vaults} />
+				<Table columns={columns} data={mockVaults} />
 			</div>
 		</div>
 	);
 }
+
+const mockVaults: DApp[] = [
+	{
+		name: "Desig Vault",
+		type: "DEX",
+		labels: ["Boosted", "Farming"],
+		apy: 11.71,
+		chain: "30.555$",
+		logo: "/assets/ui-icons/market/desigVault.svg",
+	},
+	{
+		name: "Bucket Protocol",
+		type: "LENDING",
+		labels: ["Boosted", "Farming"],
+		apy: 5.71,
+		chain: "23.555$",
+		logo: "/assets/ui-icons/market/bucketProtocol.svg",
+	},
+	{
+		name: "Navi Protocol Vault",
+		type: "LENDING",
+		labels: ["Boosted", "Farming"],
+		apy: 10.71,
+		chain: "12.215$",
+		logo: "/assets/ui-icons/market/naviProtocolValut.svg",
+	},
+	{
+		name: "Splash",
+		type: "LENDING",
+		labels: ["Boosted", "Farming"],
+		apy: 1.44,
+		chain: "20.085$",
+		logo: "/assets/ui-icons/market/splash.svg",
+	},
+	{
+		name: "Pyth Network",
+		type: "DEX",
+		labels: ["Boosted", "Farming"],
+		apy: 9.71,
+		chain: "27.715$",
+		logo: "/assets/ui-icons/market/pythNetwork.svg",
+	},
+	{
+		name: "Scallop",
+		type: "DEX",
+		labels: ["Boosted", "Farming"],
+		apy: 2.71,
+		chain: "29.885$",
+		logo: "/assets/ui-icons/market/scallop.svg",
+	},
+	{
+		name: "BlueMove",
+		type: "DEX",
+		labels: ["Boosted", "Farming"],
+		apy: 2.71,
+		chain: "11.235$",
+		logo: "/assets/ui-icons/market/blueMove.svg",
+	},
+	{
+		name: "OmniBTC",
+		type: "DEX",
+		labels: ["Boosted", "Farming"],
+		apy: 0.56,
+		chain: "00.805$",
+		logo: "/assets/ui-icons/market/omniBtc.svg",
+	},
+];
+
+const MOCK_DEPOSIT_DATA: DepositData[] = [
+	{
+		id: "totalValue",
+		title: "Total Value Routed",
+		value: "$9.62M",
+		tooltip: "",
+	},
+	{
+		id: "totalDeposit",
+		title: "Total Deposit",
+		value: "$14.72M",
+		tooltip: "",
+	},
+	{
+		id: "totalBorrow",
+		title: "Total Borrow",
+		value: "$9.23M",
+		tooltip: "",
+	},
+];
